+++ lib/toychain-data_structures/block.rb
