+++ lib/toychain-data_structures/block/header.rb
