+++ lib/toychain-data_structures/hasher.rb
