+++ lib/toychain-data_structures/message.rb
