+++ lib/toychain-data_structures/record.rb
