+++ lib/toychain-data_structures/serializer.rb
